requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
dj-database-url==2.2.0
Django==5.0.6
filelock==3.14.0
firebase-admin==6.5.0
flatbuffers==24.3.25
fsspec==2024.6.0
gast==0.5.4
git-filter-repo==2.38.0
google-api-core==2.19.0
google-api-python-client==2.132.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.1
grpcio==1.64.1
grpcio-status==1.62.2
gunicorn==22.0.0
h5py==3.11.0
httplib2==0.22.0
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
keras==3.3.3
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mkl==2021.4.0
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
namex==0.0.8
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.2
requests==2.32.3
rich==13.7.1
rsa==4.9
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
sympy==1.12.1
tbb==2021.12.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
torch==2.3.0
torchaudio==2.3.0
torchvision==0.18.0
typing_extensions==4.12.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
